+++ apps/opencrane/src/hosting/index.ts
@@ -1,3 +1,1 @@
-export type { GcpHostingConfig, HostingAdapter, TenantStateVolume, TenantStorageBinding, TenantStorageRequest } from "./hosting-adapter.types.js";
-export { HostingProvider } from "./hosting-adapter.types.js";
 export { _BuildHostingAdapter } from "./hosting-adapter.factory.js";